enabled fragment to use activityViewModel and enabled api call for list of matching exercises for playlist

+++ app/src/main/java/com/example/fitness_app_as/feature/playlist/PlaylistViewModel.kt
@@ -32,7 +32,7 @@ class PlaylistViewModel @Inject constructor(
     fun getExercisesForPlaylist(playlist: Playlist) {
         viewModelScope.launch {
             val exercises = exerciseRemoteSource.getExerciseItems()
-            val filteredExercises = exercises.filter { it.id in playlist.exerciseIds }
+            val filteredExercises = exercises.filter { it.id.toInt() in playlist.exerciseIds }
             mutablePlaylistState.emit(PlaylistState.LoadExercises(filteredExercises))
         }
     }
